refactor: move Interval to version_trait module
